Delete Cypress plugin file and add logic to Cypress config
